Add run and debug button and move code around
